+++ src/com/dev/viz/utils/Common.java
@@ -1,8 +1,8 @@
 package com.dev.viz.utils;
 
 public class Common {
-    public static int greatestNumber(int[] datas){
-        int greatestNumber = 0;
+    public static float greatestNumber(float[] datas){
+        float greatestNumber = 0;
         for(int i = 0; i < datas.length; i++){
             greatestNumber = datas[i] > greatestNumber ? datas[i] : greatestNumber ;
         }
